requirements.txt
streamlit
sqlite3
google-generativeai
python-dotenv
pandas